feat(autosave;restore-ui): restore functionality from IndexedDB
Implemented a function that allows users to restore their model progress
from the indexedDB. To do this, I've adjusted the ControlBar component
to add a "Restore Popup" button that enables users to choose the key
from the DB to restore. Additionally, the IndexedDBRestore functionality
has been revamped for completeness and effectiveness. `Deserialize`
function has been updated in the worker to parse the stringified
ModelSave correctly, facilitating the restoration process.