feat(autosave;restore-ui): restore functionality from IndexedDB; restore ui (#162)
* feat(restoreUI): add initial implemtation for restore ui

Signed-off-by: Bill ZHANG <36790218+Lutra-Fs@users.noreply.github.com>

* feat(autosave;restore-ui): restore functionality from IndexedDB

Implemented a function that allows users to restore their model progress
from the indexedDB. To do this, I've adjusted the ControlBar component
to add a "Restore Popup" button that enables users to choose the key
from the DB to restore. Additionally, the IndexedDBRestore functionality
has been revamped for completeness and effectiveness. `Deserialize`
function has been updated in the worker to parse the stringified
ModelSave correctly, facilitating the restoration process.

---------

Signed-off-by: Bill ZHANG <36790218+Lutra-Fs@users.noreply.github.com>