refactor(parameterBar): use CSS to implement hide/show function
The styling and functionality of the ParametersBar component has been
updated so it can be hidden or revealed by the user. A new styled
component, 'Pane', has been created to wrap the original 'Container'
component and added to the return statement. Transition properties have
been added to create smooth movement of the UI elements as they hide and
reveal.  The 'ShowHideButton' component has been updated to support the
new functionality. The changes aim to provide a better user interface
experience by offering more flexibility in viewing options.